my-flask-app/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
firebase-admin==6.5.0
gcsfs==2024.3.1
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.21.0
google-cloud-core==2.4.1
google-cloud-documentai==2.27.0
google-cloud-documentai-toolbox==0.13.3a0
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-cloud-vision==3.7.2
google-crc32c==1.5.0
google-resumable-media==2.7.0
google_spreadsheet==0.0.6
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.2
grpcio-status==1.62.2
gunicorn==22.0.0
pydata-google-auth==1.8.2
pygsheets==2.0.6
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
Werkzeug==3.0.3
python-dotenv==1.0.1